requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.9 